Fix harvest_entry_id integer overflow - change to BIGINT
Harvest entry IDs can exceed 2^31, causing 'integer out of range' errors.
- Add migration to alter column type to BIGINT
- Fix initial migration for fresh installs

+++ alembic/versions/20260101_0000_001_initial_schema.py
@@ -80,7 +80,7 @@ def upgrade() -> None:
         sa.Column("slack_message_ts", sa.String(30), nullable=True),
         sa.Column("slack_channel_id", sa.String(20), nullable=True),
         sa.Column("calendar_event_id", sa.String(100), nullable=True),
-        sa.Column("harvest_entry_id", sa.Integer(), nullable=True),
+        sa.Column("harvest_entry_id", sa.BigInteger(), nullable=True),
         sa.Column(
             "status",
             postgresql.ENUM(
